Rectification des packages
