--- src/Model/Cart.java
@@ -1,51 +1,0 @@
-package Model;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Cart {
-    private static Cart instance;
-    private List<CartItem> items;
-
-    private Cart() {
-        items = new ArrayList<>();
-    }
-
-    public static Cart getInstance() {
-        if (instance == null) {
-            instance = new Cart();
-        }
-        return instance;
-    }
-
-    public List<CartItem> getItems() {
-        return items;
-    }
-
-    // Ajoute le produit avec la quantité choisie. Si le produit existe déjà, on incrémente la quantité.
-    public void addProduct(Product product, int quantity) {
-        for (CartItem item : items) {
-            if (item.getProduct().getProductId() == product.getProductId()) {
-                item.setQuantity(item.getQuantity() + quantity);
-                return;
-            }
-        }
-        items.add(new CartItem(product, quantity));
-    }
-
-    public void removeProduct(Product product) {
-        items.removeIf(item -> item.getProduct().getProductId() == product.getProductId());
-    }
-
-    public void clear() {
-        items.clear();
-    }
-
-    public double getTotalPrice() {
-        double total = 0;
-        for (CartItem item : items) {
-            total += item.getProduct().getPrice() * item.getQuantity();
-        }
-        return total;
-    }
-}
